requirements.txt
aiosmtplib==2.0.1
anyio==3.6.2
autopep8==2.0.1
black==22.12.0
blinker==1.5
click==8.1.3
colorama==0.4.6
dnspython==2.3.0
email-validator==1.3.0
fastapi==0.89.1
fastapi-mail==1.2.4
fastapi-mqtt==1.0.6
gmqtt==0.6.11
greenlet==2.0.1
h11==0.14.0
httptools==0.5.0
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.2
mypy-extensions==0.4.3
paho-mqtt==1.6.1
pathspec==0.10.3
platformdirs==2.6.2
pycodestyle==2.10.0
pydantic==1.10.4
python-dotenv==0.21.0
PyYAML==6.0
sniffio==1.3.0
SQLAlchemy==1.4.46
starlette==0.22.0
tomli==2.0.1
typing_extensions==4.4.0
uvicorn==0.19.0
watchfiles==0.18.1
websockets==10.4
